backend/requirements.txt
flask
keras
Pillow
numpy
requests
tensorflow
boto3